LayoutTests/platform/ios-sim-deprecated/fast/dom/Document/CaretRangeFromPoint/caretRangeFromPoint-in-zoom-and-scroll-expected.txt

ALERT: range is null
A long text to test document caretRangeFromPoint. Test scroll, zoom, zoomAndScroll .................................... The End
PASS Test completed successfully
